chore: Rename new-game-map file
