+++ src/common/pages/new-game-page.tsx
@@ -31,7 +31,7 @@ import {
 import { ChevronDown } from "lucide-react";
 import { AddonSelect } from "@/addon/components/addon-select";
 import { api } from "@/api/api";
-import { GameMap } from "../components/map/new-game-map";
+import { GameMap } from "../components/map/game-map";
 
 export function NewGamePage() {
   const { t } = useTranslation();
